- Seperate each class from the simulation file - Brought main out as an individual file
