+++ src/Truck.py
@@ -1,0 +1,6 @@
+
+class Truck(object):
+    def __init__(self, po, size, live):
+        self.po = po
+        self.size = size
+        self.live = live
